+++ rector.php
@@ -91,6 +91,8 @@ return RectorConfig::configure()
         __DIR__ . '/tests/Unit/',
         __DIR__ . '/src/Services/Timeman',
         __DIR__ . '/tests/Integration/Services/Timeman',
+        __DIR__ . '/src/Services/IMBot',
+        __DIR__ . '/tests/Integration/Services/IMBot',
     ])
     ->withCache(cacheDirectory: __DIR__ . '/var/.cache/rector')
     ->withSets(